requirements/requirements.txt
# Paquetes de manipulación y análisis de datos
pandas>=2.0.0,<3.0.0  # Manejo de datos tabulares, creación de DataFrames

# Paquetes de visualización
matplotlib            # Gráficos y visualizaciones

# Paquetes de Machine Learning
scikit-learn>=1.3.2,<1.3.3  # Modelos de machine learning y herramientas de preprocesamiento

# Herramientas de desarrollo y pruebas
tox                   # Automatización de entornos de desarrollo y pruebas
pytest                # Framework para escribir y ejecutar pruebas
mlflow
argparse

# We use compatible release functionality (see PEP 440 here: https://www.python.org/dev/peps/pep-0440/#compatible-release)
# to specify acceptable version ranges of our project dependencies. This gives us the flexibility to keep up with small
# updates/fixes, whilst ensuring we don't install a major update which could introduce backwards incompatible changes.
numpy>=1.23.0,<2.0.0
joblib>=1.1.0,<2.0.0
pydantic>=1.10.0,<2.0.0
strictyaml>=1.7.0,<2.0.0
feature-engine>=1.6.1,<2.0.0
pyspark
pyyaml
uvicorn>=0.20.0,<0.30.0
fastapi>=0.88.0,<1.0.0
python-multipart>=0.0.5,<0.1.0
typing_extensions>=4.2.0,<5.0.0
loguru>=0.5.3,<1.0.0